git_analyst/requirements.txt
gitpython
networkx
radon
mcp
